refactor(base-driver): split basedriver helpers into a barrel
`basedriver/helpers.ts` had grown into a 580-line grab bag (app
download/caching, capability parsing, log prefixing, version lookup).
Split it into focused modules under `basedriver/helpers/`, re-exported
through `basedriver/helpers/index.ts` so existing consumers of the
`DriverHelpers`-shaped module are unaffected.

`generateDriverLogPrefix` and the base `logger` are cross-cutting
(used by `protocol/` as well as `basedriver/`), so they moved to the
shared `lib/helpers/` directory instead, removing `protocol/`'s
backward dependency on `basedriver/` internals and the resulting risk
of circular imports.
