+++ packages/base-driver/lib/basedriver/capabilities.ts
@@ -2,9 +2,9 @@ import {util} from '@appium/support';
 import type {Capabilities, Constraints, NSCapabilities, StandardCapabilities, W3CCapabilities} from '@appium/types';
 import type {KeyAsString, MergeExclusive} from 'type-fest';
 
+import {log} from '../helpers/logger.js';
 import {errors} from '../protocol/errors.js';
 import {omit, pickBy} from '../utils.js';
-import {log} from './logger.js';
 import {validator} from './validation.js';
 
 export const APPIUM_VENDOR_PREFIX = 'appium:';
