+++ packages/base-driver/lib/basedriver/helpers/index.ts
@@ -1,0 +1,4 @@
+export {configureApp} from './app-cache.js';
+export {duplicateKeys, isPackageOrBundle, parseCapsArray} from './caps.js';
+export {generateDriverLogPrefix} from '../../helpers/log-prefix.js';
+export {BASEDRIVER_VER} from './version.js';
