+++ packages/base-driver/lib/index.ts
@@ -33,8 +33,7 @@ export {DEFAULT_WS_PATHNAME_PREFIX} from './express/websocket.js';
 // BiDi exports
 export {BIDI_COMMANDS} from './protocol/bidi-commands.js';
 
-export {generateDriverLogPrefix} from './basedriver/helpers.js';
-
 export {isW3cCaps} from './helpers/capabilities.js';
+export {generateDriverLogPrefix} from './helpers/log-prefix.js';
 
 export type {ServerOpts} from './express/server.js';
