+++ packages/base-driver/test/e2e/basedriver/helpers.e2e.spec.ts
@@ -10,7 +10,7 @@ import contentDisposition from 'content-disposition';
 import finalhandler from 'finalhandler';
 import serveStatic from 'serve-static';
 
-import {configureApp} from '../../../lib/basedriver/helpers.js';
+import {configureApp} from '../../../lib/basedriver/helpers/index.js';
 
 const FIXTURE_ROOT = path.resolve(
   node.getModuleRootSync('@appium/base-driver', import.meta.filename)!,
